モバイル環境でチャンネルに投稿が適用されてなかった不具合の修正
