+++ packages/frontend/src/ui/_common_/navbar.vue
@@ -55,6 +55,7 @@
 <script lang="ts" setup>
 import { computed, defineAsyncComponent, ref, watch } from 'vue';
 import { openInstanceMenu } from './common';
+import { openPostForm } from './postnote';
 import * as os from '@/os';
 import { navbarItemDef } from '@/navbar';
 import { $i, openAccountMenu as openAccountMenu_ } from '@/account';
@@ -89,22 +90,6 @@ function openAccountMenu(ev: MouseEvent) {
 	openAccountMenu_({
 		withExtraOperation: true,
 	}, ev);
-}
-
-async function openPostForm() {
-	//チャンネルページを開いている場合はチャンネルに投稿
-	const paths = location.href.split('/');
-	let channel: any = null;
-	if (paths.length > 4) {
-		if (paths[3] === 'channels') {
-			channel = await os.api('channels/show', {
-				channelId: paths[4],
-			});
-		}
-	}
-	os.post({
-		channel: channel,
-	});
 }
 
 function more(ev: MouseEvent) {
